post,get,delete,get by id API's done

+++ src/main/java/org/example/service/impl/BookServiceImpl.java
@@ -10,6 +10,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class BookServiceImpl implements BookService {
@@ -38,8 +39,21 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public void deleteBook(Long id) {
-        repository.deleteById(id);
+    public boolean deleteBook(Long id) {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        }
+        return false;
     }
+
+    @Override
+    public Book getBookById(Long id) {
+       Optional<BookEntity> byId = repository.findById(id);
+       return mapper.map(byId,Book.class);
+
+    }
+
+
 
 }
